added plus and minus method (one bug)

+++ TimeLibrary/Helper/TimeToSecondHelper.cs
@@ -4,7 +4,7 @@
     {
         public static long Get(int hours, byte minutes = 0, byte seconds = 0)
         {
-            return (hours * TimeEnum.ONE_HOUR_IN_SECONDS) + (minutes * TimeEnum.ONE_MINUTE_IN_SECONDS) + seconds;
+            return (hours * (int)TimeEnum.ONE_HOUR_IN_SECONDS) + (minutes * (byte)TimeEnum.ONE_MINUTE_IN_SECONDS) + seconds;
         }
     }
 }
